moved corva/event/data files to proper corva/models files
