+++ corva/app/scheduled.py
@@ -1,6 +1,5 @@
 from corva.app.base import BaseApp
-from corva.models.scheduled import ScheduledContext
-from corva.event.data.scheduled import ScheduledEventData
+from corva.models.scheduled import ScheduledContext, ScheduledEventData
 from corva.event.event import Event
 from corva.event.loader.scheduled import ScheduledLoader
 from corva.state.redis_adapter import RedisAdapter
