--- corva/event/data/base.py
@@ -1,6 +1,0 @@
-from pydantic.main import BaseModel, Extra
-
-
-class BaseEventData(BaseModel):
-    class Config:
-        extra = Extra.allow
