+++ corva/event/event.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 from collections import UserList
 from typing import List
 
-from corva.event.data.base import BaseEventData
+from corva.models.base import BaseEventData
 
 
 class Event(UserList):
