+++ corva/event/loader/task.py
@@ -1,4 +1,4 @@
-from corva.event.data.task import TaskEventData
+from corva.models.task import TaskEventData
 from corva.event.event import Event
 from corva.event.loader.base import BaseLoader
