+++ corva/models/base.py
@@ -1,4 +1,4 @@
-from pydantic import BaseModel
+from pydantic import BaseModel, Extra
 
 from corva.event.event import Event
 
@@ -8,3 +8,8 @@ class BaseContext(BaseModel):
         arbitrary_types_allowed = True
 
     event: Event
+
+
+class BaseEventData(BaseModel):
+    class Config:
+        extra = Extra.allow
