+++ corva/models/task.py
@@ -1,8 +1,9 @@
 from typing import Any, Dict, Literal, Optional
 
 from pydantic import BaseModel
+from pydantic.types import conint
 
-from corva.models.base import BaseContext
+from corva.models.base import BaseContext, BaseEventData
 
 
 class TaskData(BaseModel):
@@ -25,3 +26,9 @@ class UpdateTaskInfoData(BaseModel):
 class TaskContext(BaseContext):
     task: TaskData
     save_data: dict = {}
+
+
+class TaskEventData(BaseEventData):
+    id: Optional[str] = None
+    task_id: str
+    version: conint(ge=2, le=2)  # only utils API v2 supported
