+++ tests/app/test_base.py
@@ -2,7 +2,7 @@ import pytest
 from pytest_mock import MockerFixture
 
 from corva.app.base import BaseApp
-from corva.event.data.base import BaseEventData
+from corva.models.base import BaseEventData
 from corva.event.event import Event
 from tests.conftest import ComparableException, APP_KEY, CACHE_URL
