+++ tests/app/test_scheduled.py
@@ -1,9 +1,8 @@
 import pytest
 from pytest_mock import MockerFixture
 
-from corva.models.scheduled import ScheduledContext
+from corva.models.scheduled import ScheduledContext, ScheduledEventData
 from corva.app.scheduled import ScheduledApp
-from corva.event.data.scheduled import ScheduledEventData
 from corva.event.event import Event
 from tests.conftest import APP_KEY, CACHE_URL
