+++ tests/app/test_stream.py
@@ -1,9 +1,8 @@
 import pytest
 from pytest_mock import MockerFixture
 
-from corva.models.stream import StreamContext
+from corva.models.stream import StreamContext, Record, StreamEventData
 from corva.app.stream import StreamApp
-from corva.event.data.stream import StreamEventData, Record
 from corva.event.event import Event
 from tests.conftest import APP_KEY, CACHE_URL
